A simple keyllogger.
i created this keyllogger to have the most readable output possible (that i can program of course :) ). It sends the captured characters to your email. Configure it in the msg["To"] and Password variables. 
There are still some problems, as I have to adjust the Timer and capture some numpad keys. But I'll fix it soon. I want to add a graphical interface where the user will configure his email, password and the interval between messages. Enjoy :)