DtoryFi
DtoryFi (Directory Finder) it's a basic web fuzzing tool, you can use for your pentests. Sorry if there's an error with the colors on the cmd, but they are perfect in VSCode. To use, you just need to write the address of your target, and for second the complete address of the wordlist on your computer. All the program it's in Brazillian Portuguese.