Just fixed a bug
I fixed an error where the SSID of the wifi connection was registered in the system, but without the password.  Then the program was crashing because it wasn't finding the password index in the connections list. I added a exception for when the system doesn't have the password, so the program can keep searching for others wifi connections without being interrupted.